+++ src/main/java/wooteco/subway/presentation/PathController.java
@@ -5,22 +5,23 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.application.LineService;
+import wooteco.subway.application.PathService;
 import wooteco.subway.dto.PathResponse;
 
 @RequestMapping("/paths")
 @RestController
 public class PathController {
 
-    private final LineService lineService;
+    private final PathService pathService;
 
-    public PathController(final LineService lineService) {
-        this.lineService = lineService;
+    public PathController(final PathService pathService) {
+        this.pathService = pathService;
     }
+
 
     @GetMapping
     public ResponseEntity<PathResponse> findPath(@RequestParam Long source, @RequestParam Long target, @RequestParam int age) {
-        PathResponse response = lineService.findPath(source, target, age);
+        PathResponse response = pathService.findPath(source, target, age);
         return ResponseEntity.ok(response);
     }
 }
